requirements/base.txt
Flask==2.3.2
SQLAlchemy==2.0.18
Flask-SQLAlchemy==3.0.5
Flask-RESTful==0.3.10
marshmallow==3.19.0
Flask-Migrate==4.0.4
psycopg2-binary==2.9.7
Flask-Script==2.0.6
click==8.1.4
python-dotenv==1.0.0
flasgger==0.9.7.1
apispec-webframeworks==0.5.2
loguru==0.7.0